requirements.txt
pytesseract
pillow
transformers
groq
python-dotenv
os
PyPDF2
json
re
